Changed IP Address to be local, started changes to NCAPC ID
Changed IP Address to pull local IP address of the local machine.

To support multiple clients connecting, we need to store the current clients and their information inside of either a dict or tuple, and access them based off what needs to be done. This will need to be accessed based on the client requesting them. For the initial change, we will not verify that this is a verified user, rather just respond as needed.

The Application ID will be remembered through the Discover Services initially, however, it should be noted that this is the purpose of the Join Requests. We will pull the APPL ID from the discovery request.

Current changes are in Thread132, while it does not publish to the correct new topic, it will respond back to the original topic. It also prints to the new topic to the console.

TODO: Ensure the correct topic is saved.